+++ src/lib.rs
@@ -70,7 +70,7 @@
 
 #![crate_type = "proc-macro"]
 #![recursion_limit = "256"]
-#![doc(html_root_url = "https://docs.rs/pin-project/0.1.7")]
+#![doc(html_root_url = "https://docs.rs/pin-project/0.1.8")]
 
 extern crate proc_macro;